refactor: ProjectDetails use-cases into project/section/task folders
