+++ app/src/main/java/com/example/twdist_android/di/ProjectDetailsModule.kt
@@ -8,14 +8,16 @@ import com.example.twdist_android.features.projectdetails.data.store.inmemory.In
 import com.example.twdist_android.features.projectdetails.data.store.inmemory.InMemoryTaskStore
 import com.example.twdist_android.features.explore.domain.store.ProjectStateStore
 import com.example.twdist_android.features.projectdetails.domain.store.TaskStateStore
-import com.example.twdist_android.features.projectdetails.application.usecases.GetProjectByIdUseCase
+import com.example.twdist_android.features.projectdetails.application.usecases.project.GetProjectByIdUseCase
 import com.example.twdist_android.features.projectdetails.data.repository.ProjectDetailsRepositoryImpl
 import com.example.twdist_android.features.projectdetails.domain.repository.ProjectDetailsRepository
 import com.example.twdist_android.features.projectdetails.data.repository.SectionRepositoryImpl
 import com.example.twdist_android.features.projectdetails.data.repository.TaskRepositoryImpl
 import com.example.twdist_android.features.projectdetails.domain.repository.SectionRepository
 import com.example.twdist_android.features.projectdetails.domain.repository.TaskRepository
-import com.example.twdist_android.features.projectdetails.application.usecases.*
+import com.example.twdist_android.features.projectdetails.application.usecases.project.*
+import com.example.twdist_android.features.projectdetails.application.usecases.section.*
+import com.example.twdist_android.features.projectdetails.application.usecases.task.*
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
