+++ app/src/main/java/com/example/twdist_android/features/projectdetails/application/usecases/section/RemoveTaskIdFromSectionUseCase.kt
@@ -1,13 +1,14 @@
-package com.example.twdist_android.features.projectdetails.application.usecases
+package com.example.twdist_android.features.projectdetails.application.usecases.section
 
 import com.example.twdist_android.features.projectdetails.domain.model.Section
 import com.example.twdist_android.features.projectdetails.domain.repository.SectionRepository
 import javax.inject.Inject
 
-class AddTaskIdToSectionUseCase @Inject constructor(
+class RemoveTaskIdFromSectionUseCase @Inject constructor(
     private val repository: SectionRepository
 ) {
     suspend operator fun invoke(sectionId: Long, taskId: String): Result<Section> {
-        return repository.addTaskIdToSection(sectionId, taskId)
+        return repository.removeTaskIdFromSection(sectionId, taskId)
     }
 }
+
