+++ app/src/main/java/com/example/twdist_android/features/projectdetails/application/usecases/section/ReplaceTaskIdInSectionUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.twdist_android.features.projectdetails.application.usecases
+package com.example.twdist_android.features.projectdetails.application.usecases.section
 
 import com.example.twdist_android.features.projectdetails.domain.model.Section
 import com.example.twdist_android.features.projectdetails.domain.repository.SectionRepository
@@ -15,3 +15,4 @@ class ReplaceTaskIdInSectionUseCase @Inject constructor(
         return repository.replaceTaskIdInSection(sectionId, oldTaskId, newTaskId)
     }
 }
+
