+++ app/src/main/java/com/example/twdist_android/features/projectdetails/application/usecases/section/UpdateSectionNameUseCase.kt
@@ -1,14 +1,15 @@
-package com.example.twdist_android.features.projectdetails.application.usecases
+package com.example.twdist_android.features.projectdetails.application.usecases.section
 
 import com.example.twdist_android.features.projectdetails.domain.model.Section
 import com.example.twdist_android.features.projectdetails.domain.model.SectionName
 import com.example.twdist_android.features.projectdetails.domain.repository.SectionRepository
 import javax.inject.Inject
 
-class CreateSectionUseCase @Inject constructor(
+class UpdateSectionNameUseCase @Inject constructor(
     private val repository: SectionRepository
 ) {
-    suspend operator fun invoke(projectId: Long, sectionName: SectionName): Result<Section> {
-        return repository.createSection(projectId, sectionName)
+    suspend operator fun invoke(sectionId: Long, sectionName: SectionName): Result<Section> {
+        return repository.updateSectionName(sectionId, sectionName)
     }
 }
+
